add decryption service tests
