+++ wire-ios-data-model/Source/MLS/MLSDecryptionService.swift
@@ -19,6 +19,7 @@
 import Foundation
 import WireCoreCrypto
 
+// sourcery: AutoMockable
 public protocol MLSDecryptionSerivceInterface {
 
     func decrypt(
